+++ src/framework.js
@@ -37,6 +37,7 @@ export { optimizeImage, batchDOMOperations } from './utils/dom.js';
 
 // First-party plugins
 export { iconsPlugin } from './plugins/icons/icons-plugin.js';
+export { themePlugin } from './plugins/theme/theme-plugin.js';
 
 // Import classes for internal use
 import { ComponentManager } from './core/component-manager.js';
@@ -312,5 +313,5 @@ export function createApp(config = {}) {
 }
 
 // Framework metadata
-export const FRAMEWORK_VERSION = '1.1.0';
+export const FRAMEWORK_VERSION = '1.2.0';
 export const FRAMEWORK_NAME = 'VanillaForge';